Change item weight from a fractional to an integer value
Item/ItemTemplate.Weight was a double (with a 0.1-step input on the
frontend) for no real reason — weight is always given in whole units
in this system. Changed the column, entity, DTO, and request types to
int, and dropped the fractional step from the weight inputs. Other
fractional fields (DamageModifier) stay double — that one is a genuine
multiplier (e.g. 0.5x damage on limbs), not a count, so rounding it
would break the mechanic.

+++ Witcher/Core/Wastelands.Service.Domain/Entities/Item.cs
@@ -22,7 +22,7 @@ namespace Wastelands.Service.Domain.Entities
 
 		public ItemType ItemType { get; private set; }
 
-		public double Weight { get; private set; }
+		public int Weight { get; private set; }
 
 		public int Cost { get; private set; }
